+++ src/test/java/school/redrover/OrganizationFolderTest.java
@@ -2,157 +2,169 @@ package school.redrover;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
 import org.testng.Assert;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class OrganizationFolderTest extends BaseTest {
+    private static final String PROJECT_NAME = "Organization Folder";
+    private static final String NEW_PROJECT_NAME = "Organization Folder Renamed";
 
-    private void creationNewOrganizationFolder(String folderName) {
+    private void returnHomeJenkins() {
+        getDriver().findElement(By.id("jenkins-home-link")).click();
+    }
+    private void createProject(String name) {
         getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
-        getDriver().findElement(By.id("name")).sendKeys(folderName);
-        getDriver().findElement(By.xpath("//span[text()='Organization Folder']")).click();
+        getDriver().findElement(By.name("name")).sendKeys(name);
+        getDriver().findElement(By.xpath("//li//span[text()='Organization Folder']")).click();
         getDriver().findElement(By.id("ok-button")).click();
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
+        getDriver().findElement(By.name("Submit")).click();
     }
+    private void createOrganizationFolder(String organizationFolderName) {
 
-    private void createOrganizationFolder(String name) {
-        returnToJenkinsHomePage();
-        getDriver().findElement(By.xpath("//div[@id='tasks']//a[@href='/view/all/newJob']")).click();
-        getDriver().findElement(By.xpath("//input[@id='name']")).sendKeys(name);
-        getDriver().findElement(By.xpath("//span[text()='Organization Folder']")).click();
-        getDriver().findElement(By.xpath("//div[@class='footer']//button")).click();
-    }
-
-    private void returnToJenkinsHomePage() {
-        getDriver().findElement(By.xpath("//a[@id = 'jenkins-home-link']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//input[@name='name']")).sendKeys(organizationFolderName);
+        getDriver().findElement(By.cssSelector(".jenkins_branch_OrganizationFolder")).click();
+        getDriver().findElement(By.id("ok-button")).click();
+        getDriver().findElement(By.name("Submit")).click();
+        returnHomeJenkins();
     }
 
     @Test
-    public void testCreatedNewOrganizationFolder() {
-        final String folderName = "Organization_Folder";
+    public void testCreateProject() {
+        createProject(PROJECT_NAME);
 
-        creationNewOrganizationFolder(folderName);
-        getDriver().findElement(By.linkText("Dashboard")).click();
+        Assert.assertEquals(getDriver().findElement(By.xpath("//*[@id='main-panel']/h1")).getText(),"Organization Folder");
+    }
 
-        Assert.assertTrue(getDriver()
-                .findElement(By.xpath("//tr[@id='job_" + folderName + "']")).isDisplayed());
+    @DataProvider(name = "wrong-character")
+    public Object[][] provideWrongCharacters() {
+        return new Object[][]{{"!"}, {"@"}, {"#"}, {"$"}, {"%"}, {"^"}, {"&"}, {"*"}, {"?"}, {"|"}, {">"}, {"["}, {"]"}};
+    }
 
+    @Test(dataProvider = "wrong-character")
+    public void testCreateProjectWithInvalidChar(String invalidData) {
+
+        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
+        getDriver().findElement(By.name("name")).sendKeys(invalidData);
+
+        String errorMessage = getDriver().findElement(By.id("itemname-invalid")).getText();
+
+        Assert.assertEquals(errorMessage, "» ‘" + invalidData + "’ is an unsafe character");
     }
 
     @Test
-    public void testDeleteOrganizationFolder() {
-        final String folderName = "Organization_Folder";
+    public void testCreateProjectWithSpaceInsteadOfName() {
+        final String space = " ";
 
-        creationNewOrganizationFolder(folderName);
-        getDriver().findElement(By.linkText("Dashboard")).click();
-        getDriver().findElement(By.linkText(folderName)).click();
-        getDriver().findElement(By.xpath("//span/a[@href='/job/Organization_Folder/delete']")).click();
-        getDriver().findElement(By.xpath("//button[@formnovalidate='formNoValidate']")).click();
-
-        Assert.assertEquals(getDriver().findElement(By.xpath("//h1")).getText(), "Welcome to Jenkins!");
-    }
-
-    @Test
-    public void testCreateWithValidName() {
-        final String validName = "MyOrganization";
-
-        getDriver().findElement(By.cssSelector("a[href='/view/all/newJob'")).click();
-        getDriver().findElement(By.id("name")).sendKeys(validName);
-        getDriver().findElement(By.className("jenkins_branch_OrganizationFolder")).click();
+        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
+        getDriver().findElement(By.name("name")).sendKeys(space);
+        getDriver().findElement(By.xpath("//li//span[text()='Organization Folder']")).click();
         getDriver().findElement(By.id("ok-button")).click();
 
-        getDriver().findElement(By.id("jenkins-head-icon")).click();
-
-        Assert.assertEquals(
-                getDriver().findElement(By.cssSelector("td a[href='job/" + validName + "/'] span")).getText(),
-                validName,
-                "Created organization name is incorrect");
+        Assert.assertEquals(getDriver().findElement(By.xpath("//h1")).getText(), "Error");
     }
 
     @Test
-    public void testDisableOrganizationFolder() {
+    public void testCreateProjectWithEmptyName() {
+        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//li//span[text()='Organization Folder']")).click();
 
-        creationNewOrganizationFolder("New Folder");
-
-        getDriver().findElement(By.cssSelector("button[name='Submit']")).click();
-
-        Assert.assertEquals(getDriver()
-                        .findElement(By.xpath("//button[@name='Submit']")).getText(),
-                "Enable");
+        Assert.assertEquals(getDriver().findElement(By.id("itemname-required")).getText(),
+                "» This field cannot be empty, please enter a valid name");
     }
 
     @Test
-    public void testCreateOrganizationFolderWithValidName() {
+    public void testRenameProjectFromProjectPage() {
+        createProject(PROJECT_NAME);
 
-        final String name = "Project";
+        getDriver().findElement(By.xpath("//a[contains(@href, '/confirm-rename')]")).click();
+        getDriver().findElement(By.name("newName")).clear();
+        getDriver().findElement(By.name("newName")).sendKeys(NEW_PROJECT_NAME);
+        getDriver().findElement(By.name("Submit")).click();
 
-        getDriver().findElement(By.xpath("//div[@id='tasks']//a[@href='/view/all/newJob']")).click();
-        getDriver().findElement(By.xpath("//input[@id='name']")).sendKeys(name);
-        getDriver().findElement(By.xpath("//span[text()='Organization Folder']")).click();
-        getDriver().findElement(By.xpath("//div[@class='footer']//button")).click();
-        getDriver().findElement(By.xpath("//div[@id='bottom-sticker']//button[@name='Submit']")).click();
-
-        WebElement findObject = getDriver().findElement(By.xpath("//ol[@id='breadcrumbs']/li[3]/a"));
-        String actualResult = findObject.getText();
-
-        Assert.assertEquals(actualResult, name);
+        Assert.assertEquals(getDriver().findElement(By.xpath("//h1")).getText(), NEW_PROJECT_NAME);
     }
 
     @Test
-    public void testVerifyWarningMessageEmptyName() {
+    public void testRenameProjectFromDashboardDropdownMenu() {
+        createProject(PROJECT_NAME);
 
-        final String expectedResultWarningMessage = "» This field cannot be empty, please enter a valid name";
+        getDriver().findElement(By.xpath("//a[text()='Dashboard']")).click();
+        getDriver().findElement(By.xpath("//*[@id='job_" + PROJECT_NAME + "']/td[3]/a")).click();
+        getDriver().findElement(By.linkText("Rename")).click();
+        getDriver().findElement(By.name("newName")).clear();
+        getDriver().findElement(By.name("newName")).sendKeys(NEW_PROJECT_NAME);
+        getDriver().findElement(By.name("Submit")).click();
 
-        getDriver().findElement(By.xpath("//div[@id='tasks']//a[@href='/view/all/newJob']")).click();
-        getDriver().findElement(By.xpath("//span[text()='Organization Folder']")).click();
-
-        WebElement findWarning = getDriver().findElement(By.xpath("//div[@id='itemname-required']"));
-        String actualResult = findWarning.getText();
-
-        Assert.assertEquals(actualResult, expectedResultWarningMessage);
+        Assert.assertEquals(getDriver().findElement(By.xpath("//h1")).getText(), NEW_PROJECT_NAME);
     }
 
-    @Test(dependsOnMethods = "testCreatedNewOrganizationFolder")
-    public void testOnDeletingOrganizationFolder() {
-        final String folderName = "Organization_Folder";
-        boolean deletetOK = true;
+    @Test
+    public void testRenameProjectWithSameName() {
+        createProject(PROJECT_NAME);
 
-        getDriver().findElement(By.linkText("Dashboard")).click();
-        getDriver().findElement(By.linkText(folderName)).click();
-        getDriver().findElement(By.xpath("//a[@href='/job/"+folderName+"/delete']")).click();
-        getDriver().findElement(By.xpath("//button[@formnovalidate='formNoValidate']")).click();
+        getDriver().findElement(By.xpath("//a[text()='Dashboard']")).click();
+        getDriver().findElement(By.xpath("//*[@id='job_" + PROJECT_NAME + "']/td[3]/a")).click();
+        getDriver().findElement(By.linkText("Rename")).click();
+        getDriver().findElement(By.name("Submit")).click();
 
-        try {
-            if (getDriver().findElement(By.xpath("//table[@id ='projectstatus']")).isDisplayed()) {
-                List<WebElement> elements = getDriver().findElements(By.xpath("//td/a"));
-                List<String> jobs = new ArrayList<>();
-                for (WebElement element : elements) {
-                    jobs.add(element.getText());
-                }
-                deletetOK = jobs.contains(folderName);
-            }
-        } catch (Exception e) {
-            deletetOK = false;
-        }
+        Assert.assertEquals(getDriver().findElement(By.xpath("//div//p")).getText(), "The new name is the same as the current name.");
+    }
 
-        Assert.assertFalse(deletetOK);
+    @Test
+    public void testDisableProject() {
+        createProject(PROJECT_NAME);
+
+        getDriver().findElement(By.name("Submit")).click();
+        Assert.assertEquals(getDriver().findElement(By.id("enable-project")).getText().substring(0, 46), "This Organization Folder is currently disabled");
+    }
+
+    @Test
+    public void testCloneNotExistingJob() {
+        String organizationFolderName = "Organization Folder";
+        String organizationFolderWrongName = "Organization Folder Wrong";
+        String organizationFolderCloneName = "Organization Folder Clone";
+        String errorTitle = "Error";
+        String errorMessage = "No such job:";
+
+        createOrganizationFolder(organizationFolderName);
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//input[@name='name']")).sendKeys(organizationFolderCloneName);
+        getDriver().findElement(By.xpath("//input[@id='from']")).sendKeys(organizationFolderWrongName);
+        getDriver().findElement(By.id("ok-button")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.xpath("//h1")).getText(), errorTitle);
+        Assert.assertTrue(getDriver().findElement(By.xpath("//p")).getText().contains(errorMessage));
     }
     @Test
-    public void testRedirectAfterDeleting() {
-        final String folderName = "OrganizationFolder";
+    public void testCloneOrganizationFolder() {
+        String organizationFolderName = "Organization Folder Parent";
+        String organizationFolderCloneName = "Organization Folder Clone";
+        Actions actions = new Actions(getDriver());
 
-        creationNewOrganizationFolder(folderName);
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//input[@name='name']")).sendKeys(organizationFolderName);
+        getDriver().findElement(By.cssSelector(".jenkins_branch_OrganizationFolder")).click();
+        getDriver().findElement(By.id("ok-button")).click();
+        WebElement checkBox = getDriver().findElement(By.xpath("//section[2]//label[contains(text(), 'Periodically if not otherwise run')]"));
+        actions.moveToElement(checkBox).click().build().perform();
+        getDriver().findElement(By.name("Submit")).click();
+        returnHomeJenkins();
 
-        getDriver().findElement(By.linkText("Dashboard")).click();
-        getDriver().findElement(By.linkText(folderName)).click();
-        getDriver().findElement(By.xpath("//a[@href='/job/OrganizationFolder/delete']")).click();
-        getDriver().findElement(By.xpath("//button[@formnovalidate='formNoValidate']")).click();
+        getDriver().findElement(By.xpath("//a[@href = '/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//input[@name='name']")).sendKeys(organizationFolderCloneName);
+        getDriver().findElement(By.xpath("//input[@id='from']")).sendKeys(organizationFolderName);
+        getDriver().findElement(By.id("ok-button")).click();
+        boolean isCheckBoxChecked = getDriver().findElement(By.xpath("//section[2]//input[@id=\"cb2\"]")).isSelected();
+        Assert.assertTrue(isCheckBoxChecked);
+        getDriver().findElement(By.name("Submit")).click();
+        returnHomeJenkins();
 
-        Assert.assertTrue(getDriver().getTitle().equals("Dashboard [Jenkins]"));
+        Assert.assertEquals(getDriver().findElement(By.xpath("//tr[@id='job_" + organizationFolderName + "']//td//span")).getText(), organizationFolderName);
+        Assert.assertEquals(getDriver().findElement(By.xpath("//tr[@id='job_" + organizationFolderCloneName + "']//td//span")).getText(), organizationFolderCloneName);
     }
 
     @Test
@@ -160,7 +172,7 @@ public class OrganizationFolderTest extends BaseTest {
 
         final String folderName = "OrganizationFolder";
 
-        creationNewOrganizationFolder(folderName);
+        createProject(folderName);
 
         getDriver().findElement(By.linkText("Dashboard")).click();
         getDriver().findElement(By.xpath(String.format("//td/a[@href='job/%s/']", folderName))).click();
@@ -174,7 +186,7 @@ public class OrganizationFolderTest extends BaseTest {
     public void testDisableByButton() {
         final String folderName = "OrganizationFolderEnable";
 
-        creationNewOrganizationFolder(folderName);
+        createProject(folderName);
 
         getDriver().findElement(By.linkText("Dashboard")).click();
 
@@ -184,34 +196,5 @@ public class OrganizationFolderTest extends BaseTest {
         Assert.assertEquals(getDriver().findElement(By.xpath("//button[@name='Submit']")).getText(),
                 "Enable");
         Assert.assertTrue(getDriver().findElement(By.xpath("//form[@method='post']")).getText().contains("This Organization Folder is currently disabled"));
-    }
-
-    @Test
-    public void testVerifyWarningMessageInvalidName() {
-        getDriver().findElement(By.xpath("//div[@id='tasks']//a[@href='/view/all/newJob']")).click();
-
-        List<String> specialCharacters = List.of("!", ":", "#", "*", "/", "&", "?", ";", "@", "%", "^", "[", "]",
-                "<", ">", "|", "$");
-
-        for (String element : specialCharacters) {
-            getDriver().findElement(By.xpath("//input[@id='name']")).sendKeys(element);
-            String actualResult = getDriver().findElement(By.xpath("//div[@id='itemname-invalid']")).getText();
-
-            Assert.assertEquals(actualResult, "» ‘" + element + "’ is an unsafe character");
-            getDriver().findElement(By.xpath("//input[@id='name']")).clear();
-        }
-    }
-
-    @Test
-    public void testVerifyCreatedItem() {
-        final String name = "Project";
-        
-        createOrganizationFolder(name);
-        returnToJenkinsHomePage();
-
-        Boolean actualResult = getDriver().findElement((By.xpath("//tr[@id='job_" + name + "']"))).getText().contains(name);
-
-        Assert.assertTrue(actualResult);
-
     }
 }
